Fix maxperuser (#33)
* Handle maxPerUser failure

If the maxPerUser threshold is exceeded, an exception must not be thrown.
Instead, the captcha is generated and displayed, but its expected result
isn't saved and lets it fail intentionally on form submission.

* Mark as used even if the client failed

* Display a validation error when maxPerUser is exceeded

This validation error detects a null value as expected result. This means
that the PreparerComponent didn't save the result because an application
rule failed. Actually, the single application rule is MaxRule, which
checks the maxPerUser limit.

* Delete dummy captchas

If maxPerUser is exceeded, the captchas becomes dummy and should not
clutter the DB.

* Log captcha dismiss when maxPerUser exceeded

* Add test cases

* Validate maxPerUser before creating captcha

---------

Co-authored-by: Olivier Haas <git.oshaas@dfgh.net>