Lower default value for maxPerUser, in short time
The previous default value of 1000 for maxPerUser allowed bots to do
1000 attempts. With SimpleEngine, these attempts with a random result
between 1 and 23 could success average 30 times, and had a probability of
61% to success at least one time.

To lower this possibility, the default value shoud not be higher than 100.
As a counterpart, human users musn't be deadlocked after 100 captcha loads
in a reasonable time interval. To appreciate the limit, we then restrict
the count to captchas generated less than 1 hour ago. Keep in mind that
this only affects unused or failed captchas, since successed captchas are
excluded from this count.