Enhance maxPerUser (#34)
* Set error message for notBlank validation rule

* Mark only successful captchas as used

This reverts partially b5f1d22.
We track bots sending a high number of wrong answers. Human users solving
the riddles shouldn't be affected by maxPerUsr limit.

* Remove maxPerUser validation rule

Since this feature is intended only for bots, the error message doesn't
matter.
Moreover, the inexistence of a captcha can have two meanings, either
maxPerUser or maxTime being exceeded. maxTime is intended for human users,
while maxPerUser is intended only for bots and doesn't need to be explained.

* Lower default value for maxPerUser, in short time

The previous default value of 1000 for maxPerUser allowed bots to do
1000 attempts. With SimpleEngine, these attempts with a random result
between 1 and 23 could success average 30 times, and had a probability of
61% to success at least one time.

To lower this possibility, the default value shoud not be higher than 100.
As a counterpart, human users musn't be deadlocked after 100 captcha loads
in a reasonable time interval. To appreciate the limit, we then restrict
the count to captchas generated less than 1 hour ago. Keep in mind that
this only affects unused or failed captchas, since successed captchas are
excluded from this count.

---------

Co-authored-by: Olivier Haas <git.oshaas@dfgh.net>